Separate IRR logic into generic module
